Fix wallet config loading with release build
