+++ v4/build.gradle
@@ -89,7 +89,7 @@ ext {
 
     // App dependencies
     abacusVersion = '1.13.43'
-    carteraVersion = '0.1.18'
+    carteraVersion = '0.1.19'
     //carteraVersion = '0.1.16-local-debug'    // Use this for local testing
     kollectionsVersion = '2.0.16'
