+++ v4/core/src/main/java/exchange/dydx/trading/core/CarteraSetup.kt
@@ -1,19 +1,13 @@
 package exchange.dydx.trading.core
 
-import android.content.Context
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.fragment.app.FragmentActivity
 import androidx.navigation.NavHostController
 import exchange.dydx.cartera.CarteraConfig
-import exchange.dydx.cartera.WalletConnectModalConfig
-import exchange.dydx.cartera.WalletConnectV2Config
 import exchange.dydx.cartera.WalletConnectionType
-import exchange.dydx.cartera.WalletProvidersConfig
-import exchange.dydx.cartera.WalletSegueConfig
 import exchange.dydx.cartera.walletprovider.providers.WalletConnectModalProvider
 import exchange.dydx.dydxstatemanager.AbacusStateManagerProtocol
-import exchange.dydx.trading.common.R
-import exchange.dydx.trading.core.WalletProvidersConfigUtil.getWalletProvidersConfig
+import exchange.dydx.trading.feature.workers.globalworkers.WalletProvidersConfigUtil.getWalletProvidersConfig
 import exchange.dydx.utilities.utils.Logging
 
 object CarteraSetup {
@@ -54,34 +48,5 @@ object CarteraSetup {
 
         // For debuggging
         // CarteraConfig.shared?.updateConfig(WalletProvidersConfigUtil.getWalletProvidersConfig())
-    }
-}
-
-object WalletProvidersConfigUtil {
-    fun getWalletProvidersConfig(appContext: Context, abacusStateManager: AbacusStateManagerProtocol): WalletProvidersConfig {
-        val appHostUrl = "https://" + appContext.getString(R.string.app_web_host)
-        val walletConnectV2Config = WalletConnectV2Config(
-            projectId = appContext.getString(R.string.wallet_connect_project_id),
-            clientName = appContext.getString(R.string.app_name),
-            clientDescription = appContext.getString(R.string.wallet_connect_description),
-            clientUrl = appHostUrl,
-            iconUrls = listOf<String>(appHostUrl + appContext.getString(R.string.wallet_connect_logo)),
-        )
-
-        val walletSegueConfig = WalletSegueConfig(
-            callbackUrl = appHostUrl + appContext.getString(R.string.wallet_segue_callback),
-        )
-
-        val walletIds = abacusStateManager.environment?.walletConnection?.walletConnect?.v2?.wallets?.android
-        val walletConnectModalConfig = WalletConnectModalConfig(
-            walletIds = walletIds?.toList(),
-        )
-
-        return WalletProvidersConfig(
-            walletConnectV1 = null,
-            walletConnectV2 = walletConnectV2Config,
-            walletConnectModal = walletConnectModalConfig,
-            walletSegue = walletSegueConfig,
-        )
     }
 }
