+++ v4/feature/workers/build.gradle
@@ -92,4 +92,6 @@ dependencies {
 
     implementation "org.jetbrains.kotlinx:kotlinx-serialization-json:$kotlinxSerializationVersion"
     implementation "org.jetbrains.kotlinx:kotlinx-serialization-core:$kotlinxSerializationVersion"
+
+    implementation("tz.co.asoft:kollections-interoperable:$kollectionsVersion")
 }
